Port over custom serializer tests from airflow core to task sdk
